feat:将DatapackInfoDialog分离为单独的类
